refactor: remove ElevenLabs enum namespace and global configure
Promote LogLevel, AgentState, and version to module-level API, move Events under Conversation, and replace ElevenLabs.configure with ConversationClient(logLevel:).
