+++ Sources/ElevenLabs/Internal/Utilities/EventSerializer.swift
@@ -120,7 +120,7 @@ enum EventSerializer {
         // Add source_info (equivalent to client in React Native)
         var sourceInfo: [String: Any] = [:]
         sourceInfo["source"] = "swift_sdk"
-        sourceInfo["version"] = SDKVersion.version
+        sourceInfo["version"] = version
         json["source_info"] = sourceInfo
 
         // Add user_id if provided
