+++ Sources/ElevenLabs/Public/Authorization/TokenService.swift
@@ -88,7 +88,7 @@ public struct TokenService: Sendable {
         queryItems += [
             URLQueryItem(name: "agent_id", value: agentId),
             URLQueryItem(name: "source", value: "swift_sdk"),
-            URLQueryItem(name: "version", value: SDKVersion.version)
+            URLQueryItem(name: "version", value: version)
         ]
         if let environment {
             queryItems.append(URLQueryItem(name: "environment", value: environment))
