+++ Sources/ElevenLabs/Public/Conversation/AgentState.swift
@@ -1,5 +1,15 @@
 import Foundation
 
+/// Agent state indicating what the agent is currently doing.
+public enum AgentState: Sendable, Equatable {
+    /// Agent is listening to the user
+    case listening
+    /// Agent is speaking
+    case speaking
+    /// Agent is thinking (e.g. preparing a tool call)
+    case thinking
+}
+
 /// Configuration for event-based agent state management using VAD and client events.
 /// Pass `nil` to use the default LiveKit-based behaviour.
 public struct AgentStateConfiguration: Sendable {
