+++ Tests/ElevenLabsTests/ElevenLabsTests.swift
@@ -2,12 +2,6 @@
 import XCTest
 
 final class ElevenLabsTests: XCTestCase {
-    func testConfigurationDefault() {
-        let config = ElevenLabs.Configuration.default
-        XCTAssertEqual(config.logLevel, .warning)
-        XCTAssertFalse(config.debugMode)
-    }
-
     func testConversationConfigInit() {
         let config = ConversationConfig()
         XCTAssertNil(config.agentOverrides)
